feat: Implement BottomNavigationBar
Implemented BottomNavigationBar and made it to be shown up when 'GO'
button is tapped. Also, created three pages for webview, recommended
page, and mypage. Made all the buttons to show each pages and used
IndexedStack for preventing pages keep refreshing when other pages are
shown up.

Pushed on the branch 'issue-7' by mistake so pushing again on
'issue-11'.