refactor: Refactor main and sub text fields
Deleted double type value but instead added SizedBox widget so that the
texts to be shown the same but also simplify and lightened the codes.
Also, instead of putting 'image No.n' on the ImageSlider card, added the
cafe's name on it by getting the info from the api url.