Feat: 회원탈퇴 모달창 구현 (#17)
* feat: Implement route transition & webview code

Route transition animation has changed from default to slide (right to left) animation. This animation happens when the 'GO' button at the bottom is pressed. And webview code is completed either.

* feat: Implement route transition & webview code

Route transition animation has changed from default to slide (right to left) animation. This animation happens when the 'GO' button at the bottom is pressed. And webview code is completed either.

* feat: Implement route transition & webview code

Route transition animation has changed from default to slide (right to left) animation. This animation happens when the 'GO' button at the bottom is pressed. And webview code is completed either.

* style: Categorize files by class

Created 'myhomepage.dart' and 'sliderightroute.dart' files for better
readability and classification. And also deleted the unnecessary annotations.

* feat: File merging & Implementing two way channel

Merged flutter files and now when the 'go' button pressed, it shows up
the webview page. Implemented two way channel between webveiw and
javascript. Not sure if it works well for now so need to keep update
with this function.

* fix: Added unadded files

Added some files that were missing after merging

* fix: Solve conflicts

Solved with some conflicts occured by merging from 'pubspec.lock'

* fix: Solve conflicts

Solved with some conflicts occured by merging from 'pubspec.lock'

* feat : flutter - 웹뷰 JS 채널 연동 기능 구현

* feat : Implement webview JS two way channel

Implemented two way channel between webview and JS.

* feat : Implement webview JS channel

Implemented channel between webview and javascript. Not yet sure if it
works well so needed to be inspected.

* feat : 내 위치 찾기 기능 구현

* feat: Implement BottomNavigationBar

Implemented BottomNavigationBar and made it to be shown up when 'GO'
button is tapped. Also, created three pages for webview, recommended
page, and mypage. Made all the buttons to show each pages and used
IndexedStack for preventing pages keep refreshing when other pages are
shown up.

* feat: Implement BottomNavigationBar

Implemented BottomNavigationBar and made it to be shown up when 'GO'
button is tapped. Also, created three pages for webview, recommended
page, and mypage. Made all the buttons to show each pages and used
IndexedStack for preventing pages keep refreshing when other pages are
shown up.

* feat: Implement BottomNavigationBar

Implemented BottomNavigationBar and made it to be shown up when 'GO'
button is tapped. Also, created three pages for webview, recommended
page, and mypage. Made all the buttons to show each pages and used
IndexedStack for preventing pages keep refreshing when other pages are
shown up.

Pushed on the branch 'issue-7' by mistake so pushing again on
'issue-11'.

* feat: Implement recommend page

Implemented recommend page but still need to connect the API so that
it can get show up the images and names of the cafes on each cards.

* feat: Implement modal screen

Implemented modal screen and it pops up when the 'mypage' button from
the BottomNavigationBar is clicked. Also implemented a bit of API
connection on 'recommend' page but still need to take a look more.

* feat: Put ripple effect on login buttons

Editted two login buttons, 'via_apple' and 'via_google', by using
'Material' and 'Ink.image'.

* fix: Change compileSdkVersion from 33 to 32

Changed the compileSdkVersion from 33 to 32. You can check it in the
'build.gradle' file.

* feat: Implement account deleting modal screen

Implemented a modal screen for deleting account.

* fix: Changed the way of "delete_account" pops up

Used the 'Navigator' function at the first and found out that it makes
the background totally black when I tap on the '회원탈퇴' button. So
changed the 'Navigator' to 'showDialog' which is a right way, and now
the background is still showing up without being totally black.

---------

Co-authored-by: Sangjun-man <molung@icloud.com>
Co-authored-by: SangJun Lee <66112027+Sangjun-man@users.noreply.github.com>