Feat: 액세스토큰 재발급기능, 즐겨찾기 조회, 수정 기능 구현 (#25)
* feat: Implement route transition & webview code

Route transition animation has changed from default to slide (right to left) animation. This animation happens when the 'GO' button at the bottom is pressed. And webview code is completed either.

* feat: Implement route transition & webview code

Route transition animation has changed from default to slide (right to left) animation. This animation happens when the 'GO' button at the bottom is pressed. And webview code is completed either.

* feat: Implement route transition & webview code

Route transition animation has changed from default to slide (right to left) animation. This animation happens when the 'GO' button at the bottom is pressed. And webview code is completed either.

* style: Categorize files by class

Created 'myhomepage.dart' and 'sliderightroute.dart' files for better
readability and classification. And also deleted the unnecessary annotations.

* feat: File merging & Implementing two way channel

Merged flutter files and now when the 'go' button pressed, it shows up
the webview page. Implemented two way channel between webveiw and
javascript. Not sure if it works well for now so need to keep update
with this function.

* fix: Added unadded files

Added some files that were missing after merging

* fix: Solve conflicts

Solved with some conflicts occured by merging from 'pubspec.lock'

* fix: Solve conflicts

Solved with some conflicts occured by merging from 'pubspec.lock'

* feat : flutter - 웹뷰 JS 채널 연동 기능 구현

* feat : Implement webview JS two way channel

Implemented two way channel between webview and JS.

* feat : Implement webview JS channel

Implemented channel between webview and javascript. Not yet sure if it
works well so needed to be inspected.

* feat : 내 위치 찾기 기능 구현

* feat: Implement BottomNavigationBar

Implemented BottomNavigationBar and made it to be shown up when 'GO'
button is tapped. Also, created three pages for webview, recommended
page, and mypage. Made all the buttons to show each pages and used
IndexedStack for preventing pages keep refreshing when other pages are
shown up.

* feat: Implement BottomNavigationBar

Implemented BottomNavigationBar and made it to be shown up when 'GO'
button is tapped. Also, created three pages for webview, recommended
page, and mypage. Made all the buttons to show each pages and used
IndexedStack for preventing pages keep refreshing when other pages are
shown up.

* feat: Implement BottomNavigationBar

Implemented BottomNavigationBar and made it to be shown up when 'GO'
button is tapped. Also, created three pages for webview, recommended
page, and mypage. Made all the buttons to show each pages and used
IndexedStack for preventing pages keep refreshing when other pages are
shown up.

Pushed on the branch 'issue-7' by mistake so pushing again on
'issue-11'.

* feat: Implement recommend page

Implemented recommend page but still need to connect the API so that
it can get show up the images and names of the cafes on each cards.

* feat: Implement modal screen

Implemented modal screen and it pops up when the 'mypage' button from
the BottomNavigationBar is clicked. Also implemented a bit of API
connection on 'recommend' page but still need to take a look more.

* feat: Put ripple effect on login buttons

Editted two login buttons, 'via_apple' and 'via_google', by using
'Material' and 'Ink.image'.

* fix: Change compileSdkVersion from 33 to 32

Changed the compileSdkVersion from 33 to 32. You can check it in the
'build.gradle' file.

* Fix : unused file delete

* feat: Implement account deleting modal screen

Implemented a modal screen for deleting account.

* fix: Changed the way of "delete_account" pops up

Used the 'Navigator' function at the first and found out that it makes
the background totally black when I tap on the '회원탈퇴' button. So
changed the 'Navigator' to 'showDialog' which is a right way, and now
the background is still showing up without being totally black.

* feat: Implement splash screen

Implemented a splash screen. Also, to put some animations like fade in
and fade out, editted the 'main.dart' file.

* feat: Implement api connection and fetch image

Implemented these features.
1. Display the 'request location permission' pop up page when tapping the
   recommended page.
2. Getting location information and put the values on api urls.
3. Implementing these pages. (Popular cafes will be implemented later)
	3-1. Close cafes from the current location
	3-2. High rated cafes
	3-3. Newly opened cafes
4. Simplifiy the "ImageSlider"

* feat: add flutter_web_auth pub, test google login

* feat: custom url scheme  설정(hipspot-mobile)

* refactor: Refactor api related codes

Refactored codes that call api. Now, it only call the api just once and
who the images through the urls that it got. Also, added some enum files
so that it doesn't pass hardcoded values as parameters.

* feat: 구글 로그인 진입지점 Url 수정

* rebase

* feat: 마이페이지 Favorite 컴포넌트 분리

* refactor: 온보딩 페이지, 메인 페이지 컴포넌트 및 폴더구조 리팩토링

key changes
  - screen/pages 폴더 삭제
    screen이라는 폴더 자체가 page의 의미 또한 포함되어있다고 생각하여, pages 폴더에 있는 파일들을 screen 폴더로 이동시켰습니다.
  - Onboarding 컴포넌트들 component/Onboarding 폴더에 분리
    onboarding 페이지 이외에 사용되지 않는 컴포넌트들이라 판단하여, Onboarding 폴더안에 모아두어 관리하고자 하였습니다.
  - Gobutton에 onTap 파라미터 추가
    onboarding_screen에서 go 버튼을 눌렀을 때, 어느 페이지로 이동하는지 확실히 알 수 있도록 DI를 통해 onTap 파라미터를 추가하였습니다.

* feat: 마이페이지 접근 시 로그인 확인하는 로직 추가

* refactor: Refactor main and sub text fields

Deleted double type value but instead added SizedBox widget so that the
texts to be shown the same but also simplify and lightened the codes.
Also, instead of putting 'image No.n' on the ImageSlider card, added the
cafe's name on it by getting the info from the api url.

* refactor: 메인스크린에 보여지는 페이지 마운트 로직 수정

- webView로 띄워지는 페이지는 마운트 이후 수정 x, 이외 페이지는 접근시 마다 새로 마운트하도록 코드를 수정했습니다

* chore: dio 패키지 추가

* feat: dio 패키지에 redirect 추가 임시커밋

* feat: 로그인 이후 favorite 요청 기능 구현

* feat: 마이페이지 즐겨찾기 목록 토글 기능 구현

* feat: 마이페이지 유저 조회, 탈퇴 기능 구현

* feat: 리프레시토큰을 통한 엑세스토큰 재발급 이후 api 재요청 기능 구현

- 엑세스토큰 재발급 과정
  엑세스토큰이 만료되면 서버는 리프레시토큰 유효성을 확인한뒤, 유효성이 확인된 경우
  statusCode 700번으로 body에 엑세스토큰, 리프레시토큰 전달하도록 설계

- 재발급받은 엑세스토큰으로 api 재요청 과정
  dio에서 statusCode 700번의 경우 에러로 처리, 따라서 interceptor onError 메서드에서
  statusCode 700번 응답에 대해, 토큰 저장 및 기존 요청 url에 새로운 accessToken으로
  요청하도록 처리

* refactor : dio 관련 코드 리팩토링

- TokenStorage 클래스 작성
  엑세스토큰, 리프레시토큰 값의 crud를 담당하는 클래스를 작성하여, 로직과 데이터 간 결합도를
  낮추고자 했습니다
- Interceptor 코드 주석 작성
  로직 파악을 쉽게 하기 위해서 적절하게 추상화하고, 주석을 추가하였습니다.
- api 폴더 위치 변경
  libs/utils/api -> libs/api 로 폴더 위치를 변경하였습니다

* refactor: 엑세스토큰 재발급과정에서 보이는 print() 내용 수정

* refactor : token 키값 하등코딩 된 부분 const에 따로 작성

* refactor : 사용하지 않는 패키지 Import 제거, 변수 할당 제거

* fix: pubspec 컨플릭트 처리

* refactor: 재발급 로직 확인 편의를 위한 수정사항 적용

---------

Co-authored-by: unknown <world2222222222@gmail.com>
Co-authored-by: yeoularu <yeoularu@gmail.com>