Feat: Getting rate and distance info
Now on the recommend page, you can not only get the info with the name
of the cafe but also the rate and distance from current location. Also,
refactored codes from 'Recommend.dart'. Now it gets each lists that it
got from api url as parameter. Lastly, refactored a bit from
'recommend_title_text_enum.dart'. Changed all the 'Text' and 'SubText'
to 'Title' and 'SubTitle' for better recognition.