Feat: Sending selected filter to web
Implemented a function to send which ID(hipspot, study, reasonable etc)
is selected when pressing the 'GO' button. Available to check this code
from 'onboarding_screen.dart' file and onTap part of GoButton.

Also, fixed a bug that happens to show the users the recommend page
right on the map. To fix this issue, wrapped Visibility widget from
'recommend_screen.dart' with Container widget and set basic color as
white.