Feat: 첫 화면에서 필터 선택 후 GO 버튼 누를 시 필터 적용된 map 바로 띄우기 (#36)
* Feat: Getting rate and distance info

Now on the recommend page, you can not only get the info with the name
of the cafe but also the rate and distance from current location. Also,
refactored codes from 'Recommend.dart'. Now it gets each lists that it
got from api url as parameter. Lastly, refactored a bit from
'recommend_title_text_enum.dart'. Changed all the 'Text' and 'SubText'
to 'Title' and 'SubTitle' for better recognition.

* Feat: Getting rate and distance info

Now on the recommend page, you can not only get the info with the name
of the cafe but also the rate and distance from current location. Also,
refactored codes from 'Recommend.dart'. Now it gets each lists that it
got from api url as parameter. Lastly, refactored a bit from
'recommend_title_text_enum.dart'. Changed all the 'Text' and 'SubText'
to 'Title' and 'SubTitle' for better recognition.

* Feat: Sending selected filter to web

Implemented a function to send which ID(hipspot, study, reasonable etc)
is selected when pressing the 'GO' button. Available to check this code
from 'onboarding_screen.dart' file and onTap part of GoButton.

Also, fixed a bug that happens to show the users the recommend page
right on the map. To fix this issue, wrapped Visibility widget from
'recommend_screen.dart' with Container widget and set basic color as
white.

* feat: Getting request and sending selectedFilter

Implemented the way of getting request for the value of the
selectedFilter and sending the value back to the web. Now, it shows the
result with the filter selected on the map successfully.

* refactor: Setting back the webview url

Set the url in webview_screen.dart back as 'https://hipspot.xyz/'.

* feat: 웹뷰에서 보낸 메세지가 유효하지 않은 경우 에러 던지지 않음

* refactor: 변수명 lowerCamelCase로 변경

---------

Co-authored-by: yeoularu <yeoularu@gmail.com>